establishing baseline on USCORE 3.0.0
eLTSS IG has been refactored to be run through FSH and QA was done to repair syntax that is no longer considered valid.

eLTSS IG has been reconfigured to be able to run with the current IG Publisher tooling.
JIRA tickets have been assessed and work on them is commencing.