FHIR-24902
Resolution Description:
This an HL7 tool rendering issue to be addressed by HL7.

Description
In the RiskAssessment profile (http://hl7.org/fhir/us/eltss/RiskAssessment-eltss.html) of the eLTSS IG, RiskAssessment.prediction.outcome appears to be bound to a value set incorrectly. It currently reads Binding: (unbound) (example). The "(unbound)" should probably be a pointer to the example value set or the attempted binding should be removed altogether (as it is in the base resource).